+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 * Move `Data.Sequences.Lazy` to `mono-traversable` 1.0.0
 * Move `Data.Textual.Encoding` to `mono-traversable` 1.0.0
+* Switch from blaze-builder to bytestring 0.10.2 or later
 
 ## 0.2.0